backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
neo4j==4.4.0
python-dotenv==0.19.0
flask-jwt-extended==4.3.1
bcrypt==3.2.0
python-jose==3.3.0 